+++ pkg/compactor/compactor.go
@@ -13,6 +13,7 @@ import (
 	"github.com/go-kit/kit/log/level"
 	"github.com/pkg/errors"
 	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/promauto"
 	"github.com/prometheus/prometheus/tsdb"
 	"github.com/thanos-io/thanos/pkg/block"
 	"github.com/thanos-io/thanos/pkg/compact"
@@ -34,6 +35,7 @@ type Config struct {
 	DataDir              string                   `yaml:"data_dir"`
 	CompactionInterval   time.Duration            `yaml:"compaction_interval"`
 	CompactionRetries    int                      `yaml:"compaction_retries"`
+	DeletionDelay        time.Duration            `yaml:"deletion_delay"`
 
 	// Compactors sharding.
 	ShardingEnabled bool       `yaml:"sharding_enabled"`
@@ -62,6 +64,10 @@ func (cfg *Config) RegisterFlags(f *flag.FlagSet) {
 	f.DurationVar(&cfg.CompactionInterval, "compactor.compaction-interval", time.Hour, "The frequency at which the compaction runs")
 	f.IntVar(&cfg.CompactionRetries, "compactor.compaction-retries", 3, "How many times to retry a failed compaction during a single compaction interval")
 	f.BoolVar(&cfg.ShardingEnabled, "compactor.sharding-enabled", false, "Shard tenants across multiple compactor instances. Sharding is required if you run multiple compactor instances, in order to coordinate compactions and avoid race conditions leading to the same tenant blocks simultaneously compacted by different instances.")
+	f.DurationVar(&cfg.DeletionDelay, "compactor.deletion-delay", 48*time.Hour, "Time before a block marked for deletion is deleted from bucket. "+
+		"If not 0, blocks will be marked for deletion and compactor component will delete blocks marked for deletion from the bucket. "+
+		"If delete-delay is 0, blocks will be deleted straight away. Note that deleting blocks immediately can cause query failures, "+
+		"if store gateway still has the block loaded, or compactor is ignoring the deletion because it's compacting the block at the same time.")
 }
 
 // Compactor is a multi-tenant TSDB blocks compactor based on Thanos.
@@ -94,6 +100,10 @@ type Compactor struct {
 	compactionRunsStarted   prometheus.Counter
 	compactionRunsCompleted prometheus.Counter
 	compactionRunsFailed    prometheus.Counter
+
+	blocksCleaned           prometheus.Counter
+	blockCleanupFailures    prometheus.Counter
+	blocksMarkedForDeletion prometheus.Counter
 
 	// TSDB syncer metrics
 	syncerMetrics *syncerMetrics
@@ -137,23 +147,31 @@ func newCompactor(
 		syncerMetrics:                      newSyncerMetrics(registerer),
 		createBucketClientAndTsdbCompactor: createBucketClientAndTsdbCompactor,
 
-		compactionRunsStarted: prometheus.NewCounter(prometheus.CounterOpts{
+		compactionRunsStarted: promauto.With(registerer).NewCounter(prometheus.CounterOpts{
 			Name: "cortex_compactor_runs_started_total",
 			Help: "Total number of compaction runs started.",
 		}),
-		compactionRunsCompleted: prometheus.NewCounter(prometheus.CounterOpts{
+		compactionRunsCompleted: promauto.With(registerer).NewCounter(prometheus.CounterOpts{
 			Name: "cortex_compactor_runs_completed_total",
 			Help: "Total number of compaction runs successfully completed.",
 		}),
-		compactionRunsFailed: prometheus.NewCounter(prometheus.CounterOpts{
+		compactionRunsFailed: promauto.With(registerer).NewCounter(prometheus.CounterOpts{
 			Name: "cortex_compactor_runs_failed_total",
 			Help: "Total number of compaction runs failed.",
 		}),
-	}
 
-	// Register metrics.
-	if registerer != nil {
-		registerer.MustRegister(c.compactionRunsStarted, c.compactionRunsCompleted, c.compactionRunsFailed)
+		blocksCleaned: promauto.With(registerer).NewCounter(prometheus.CounterOpts{
+			Name: "cortex_compactor_blocks_cleaned_total",
+			Help: "Total number of blocks deleted in compactor.",
+		}),
+		blockCleanupFailures: promauto.With(registerer).NewCounter(prometheus.CounterOpts{
+			Name: "cortex_compactor_block_cleanup_failures_total",
+			Help: "Failures encountered while deleting blocks in compactor.",
+		}),
+		blocksMarkedForDeletion: promauto.With(registerer).NewCounter(prometheus.CounterOpts{
+			Name: "cortex_compactor_blocks_marked_for_deletion_total",
+			Help: "Total number of blocks marked for deletion in compactor.",
+		}),
 	}
 
 	c.Service = services.NewBasicService(c.starting, c.running, c.stopping)
@@ -311,6 +329,10 @@ func (c *Compactor) compactUser(ctx context.Context, userID string) error {
 	// blocks that fully submatches the source blocks of the older blocks.
 	deduplicateBlocksFilter := block.NewDeduplicateFilter()
 
+	// While fetching blocks, we filter out blocks that were marked for deletion by using IgnoreDeletionMarkFilter.
+	// The delay of deleteDelay/2 is added to ensure we fetch blocks that are meant to be deleted but do not have a replacement yet.
+	ignoreDeletionMarkFilter := block.NewIgnoreDeletionMarkFilter(ulogger, bucket, time.Duration(c.compactorCfg.DeletionDelay.Seconds()/2)*time.Second)
+
 	fetcher, err := block.NewMetaFetcher(
 		ulogger,
 		c.compactorCfg.MetaSyncConcurrency,
@@ -320,9 +342,12 @@ func (c *Compactor) compactUser(ctx context.Context, userID string) error {
 		// the directory used by the Thanos Syncer, whatever is the user ID.
 		path.Join(c.compactorCfg.DataDir, "meta-"+userID),
 		reg,
-		// List of filters to apply (order matters).
-		block.NewConsistencyDelayMetaFilter(ulogger, c.compactorCfg.ConsistencyDelay, reg).Filter,
-		deduplicateBlocksFilter.Filter,
+		[]block.MetadataFilter{
+			// List of filters to apply (order matters).
+			block.NewConsistencyDelayMetaFilter(ulogger, c.compactorCfg.ConsistencyDelay, reg),
+			ignoreDeletionMarkFilter,
+			deduplicateBlocksFilter,
+		},
 	)
 	if err != nil {
 		return err
@@ -334,6 +359,8 @@ func (c *Compactor) compactUser(ctx context.Context, userID string) error {
 		bucket,
 		fetcher,
 		deduplicateBlocksFilter,
+		ignoreDeletionMarkFilter,
+		c.blocksMarkedForDeletion,
 		c.compactorCfg.BlockSyncConcurrency,
 		false, // Do not accept malformed indexes
 		true,  // Enable vertical compaction
@@ -357,7 +384,17 @@ func (c *Compactor) compactUser(ctx context.Context, userID string) error {
 		return errors.Wrap(err, "failed to create bucket compactor")
 	}
 
-	return compactor.Compact(ctx)
+	if err := compactor.Compact(ctx); err != nil {
+		return errors.Wrap(err, "compaction")
+	}
+
+	blocksCleaner := compact.NewBlocksCleaner(ulogger, bucket, ignoreDeletionMarkFilter, c.compactorCfg.DeletionDelay, c.blocksCleaned, c.blockCleanupFailures)
+
+	if err := blocksCleaner.DeleteMarkedBlocks(ctx); err != nil {
+		return errors.Wrap(err, "error cleaning blocks")
+	}
+
+	return nil
 }
 
 func (c *Compactor) discoverUsers(ctx context.Context) ([]string, error) {