+++ pkg/storage/tsdb/bucket_client_mock.go
@@ -10,6 +10,8 @@ import (
 	"github.com/stretchr/testify/mock"
 )
 
+var errObjectDoesNotExist = errors.New("object does not exist")
+
 // BucketClientMock mocks objstore.Bucket
 type BucketClientMock struct {
 	mock.Mock
@@ -54,7 +56,11 @@ func (m *BucketClientMock) MockIter(prefix string, objects []string, err error) 
 // Get mocks objstore.Bucket.Get()
 func (m *BucketClientMock) Get(ctx context.Context, name string) (io.ReadCloser, error) {
 	args := m.Called(ctx, name)
-	return args.Get(0).(io.ReadCloser), args.Error(1)
+	val, err := args.Get(0), args.Error(1)
+	if val == nil {
+		return nil, err
+	}
+	return val.(io.ReadCloser), err
 }
 
 // MockGet is a convenient method to mock Get() and Exists()
@@ -64,8 +70,12 @@ func (m *BucketClientMock) MockGet(name, content string, err error) {
 		m.On("Get", mock.Anything, name).Return(ioutil.NopCloser(bytes.NewReader([]byte(content))), err)
 	} else {
 		m.On("Exists", mock.Anything, name).Return(false, err)
-		m.On("Get", mock.Anything, name).Return(nil, errors.New("object does not exist"))
+		m.On("Get", mock.Anything, name).Return(nil, errObjectDoesNotExist)
 	}
+}
+
+func (m *BucketClientMock) MockDelete(name string, err error) {
+	m.On("Delete", mock.Anything, name).Return(err)
 }
 
 // GetRange mocks objstore.Bucket.GetRange()
@@ -82,12 +92,7 @@ func (m *BucketClientMock) Exists(ctx context.Context, name string) (bool, error
 
 // IsObjNotFoundErr mocks objstore.Bucket.IsObjNotFoundErr()
 func (m *BucketClientMock) IsObjNotFoundErr(err error) bool {
-	if err == nil {
-		return false
-	}
-
-	args := m.Called(err)
-	return args.Bool(0)
+	return err == errObjectDoesNotExist
 }
 
 // ObjectSize mocks objstore.Bucket.ObjectSize()